--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,6 +1,6 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/N1shko/otus-golang/hw10_program_optimization
 
-go 1.22
+go 1.23.6
 
 require github.com/stretchr/testify v1.7.0